Add public API for the semanticdb package and use it in SemanticdbTextDocumentProvider
[Cherry-picked a0cfa40d242b19ae0abcb51161f200c06c4cf4ec][modified]
