+++ compiler/src/dotty/tools/dotc/Compiler.scala
@@ -10,7 +10,6 @@ import Phases.Phase
 import transform.*
 import backend.jvm.GenBCode
 import localopt.StringInterpolatorOpt
-import semanticdb.ExtractSemanticDB.{ExtractSemanticInfo, AppendDiagnostics as AppendSemanticDiagnostics}
 
 /** The central class of the dotc compiler. The job of a compiler is to create
  *  runs, which process given `phases` in a given `rootContext`.
@@ -40,7 +39,7 @@ class Compiler {
          CheckShadowing()) ::       // Check for shadowed elements
     List(new YCheckPositions) ::    // YCheck positions
     List(new sbt.ExtractDependencies) :: // Sends information on classes' dependencies to sbt via callbacks
-    List(ExtractSemanticInfo()) ::  // Extract info into .semanticdb files
+    List(new semanticdb.ExtractSemanticInfo) :: // Extract info into .semanticdb files
     List(new PostTyper) ::          // Additional checks and cleanups after type checking
     List(new sjs.PrepJSInterop) ::  // Additional checks and transformations for Scala.js (Scala.js only)
     List(new sbt.ExtractAPI) ::     // Sends a representation of the API of classes to sbt via callbacks
@@ -87,7 +86,7 @@ class Compiler {
     List(new CheckCaptures.Pre) ::   // Preparations for check captures phase, enabled under captureChecking
     List(new CheckCaptures) ::       // Check captures, enabled under captureChecking
     List(CheckUnused.PostPatMat()) :: // Check for unused elements and report
-    List(AppendSemanticDiagnostics()) :: // Attach warnings to extracted SemanticDB and write to .semanticdb file
+    List(new semanticdb.AppendDiagnostics) :: // Attach warnings to extracted SemanticDB and write to .semanticdb file
     List(new ElimOpaque,             // Turn opaque into normal aliases
          new sjs.ExplicitJSClasses,  // Make all JS classes explicit (Scala.js only)
          new ExplicitOuter,          // Add accessors to outer classes from nested ones.
